initial commit of devops
The end goal is a completely automated setup of our software. This can then be
used for real on-the-metal deloyment, cloud deployment and automated testing.

Currently only setup of cpauth is implemented and is about 80% complete. On
executing "vagrant up" a VirtualBox VM is created containing CentOS 7 and inside
this VM the following happens:
  + nginx is installed and configured
  + mongodb and restheart are installed and enabled
  + cpauth is installed
    + a separate cpauth user is created
    + the cpauth repository is cloned from github
	+ scala / sbt is installed
	+ cpauth is compiled using 'sbt assembly'
	+ a systemd service files is installed that runs cpauth

What's missing to complete cpauth support is:
  + clean up the ansible files
  + check that the restheart initial database is properly created
  + execute external integration test (i.e a python script that acts as a web
    browser and uses the VM to authenticate with cpauth using e.g SAML)

After that it's a matter of adding the meta and data repositories. Part of the
work is already done (setting up CentOS etc).