[devops] Unflakeyfied the jarservice role
Include servicename in task names.

When copying the jarfile to the remote machine, save the result so that the
destination filename is accessible when creating the symbolic link.

When creating the symbolic link, include the 'state: link' field (apparently the
default behaviour has changed since I wrote that task).

Move the "enable systemd {{ servicename }}" task outside the block. Since the
block is protected with a 'when' we run into problems when ansible crashes
before the block protected by a 'when' but _after_ the task that registers the
'when'.