Use tags in the toplevel stilt playbook.
An earlier attempt was made to tag the stiltweb role, by tagging the
tasks. However the ansible documentatio - although somewhat uncoherent - states
that "In order for tags to be applied to a role and all its dependencies, the
tag should be applied to the role, not to all the tasks within a role.".

Thus a role should always be tagged at toplevel if one wants its dependencies to
run - which we do.