Only use ansible tags in toplevel playbook
This seems to be the only reliable way to get the expected behaviour of
including dependencies when specifying tags in newer versions of ansible.

This means that when provisioning e.g cpmeta, instead of doing:
  ansible-playbook icosprod.yml -t "setup,linux,docker,ansible,cpmeta'

It's enough to:
  ansible-playbook icosprod.yml -t cpmeta

Previous to this patch we often used a hack wherein a roles's task/main.yml file
would import the actual tasks while tagging them all at once. This however, did
not extend to the role's dependencies.