Remove the cp*_jar_file variables from vars.yml
The intent is to make it harder to mistakenly deploy old jar files.

Example:
  + In vars.yml, stiltcluster_jar_file = stiltcluster-assembly-0.1.0.jar
  + The stiltcluster version number is increased to 0.2.0 (true story!) and
  recompiled (but the old jar file still exists)
  + The sbt 'cpDeploy' command is used to deploy the 0.2.0 version and all is
  fine.
  + Sometimes later someone (*cough*) decides to re-deploy stiltcluster and
  issue an ansible comman from the infrastruture repo. The vars.yml file is
  consulted and decides to deploy the (old) 0.1.0 version.
  + Now we've deployed an old version to production.

This patch removes the *_jar_file from vars.yml, this will make ansible fail
unless those variables are provided on the command line.

Going forward we'll only deploy jarfile services from within sbt using the
'cpDeploy' command - which does so by providing the *_jar_file variable on the
command line.

is that when the jar file version is bumbed in build.sbt tFor example,
the vars.yml have 'stiltcluster_jar_file' pointing to
stiltcluster-assembly-0.1.0.jar - but that's the old version!