icos.geoip - a sqlite backed geolocation cache
This will deploy geoip, an HTTP-based sqlite3-backed wrapper for the ipstack.com
geolocation service. It takes an IP number as input and responds with longitude
and latitude.

Requests to to https://geoip.icos-cp.eu/ips/IPNUMBER are forwarded to
ipstack.com (using our API key). The results is cached in a sqlite database and
returned to the requestor.

Note that requests to geoip.icos-cp.eu are 1) https only 2) restricted by
ansible variable {{ geoip_nginx_allow_deny }}.