icos.stiltweb - add systemd services for the stiltweb ssh tunnels
The individual tunnels are systemd instantiated services named
stiltweb_tunnel@HOST.service, e.g stiltweb_tunnel@centos.service. The tunnels
are run as the stiltweb user and the hosts are found in ~stiltweb/.ssh/config.

The tunnels are PartOf stiltweb_tunnels.targets which is a wanted by the
stiltweb.service.

Every host find in the inventory group 'stiltcluster' gets an SSH tunnel,
_unless_ it's the same host as stiltweb.

When deploying and re-deploying, we'll attempt to stop all tunnels that go to
hosts no longer in the stiltcluster group and to add new ones.