stilt - remove fsicos2 from the stiltcluster group use the role instead
Since fsicos2 is used both as the stiltweb host and as one of the hosts in the
cluster, some of the dependency roles (linux, nginx, docker, etc) ran twice for
fsicos2. By removing fsicos2 from the stiltcluster group and instead adding the
stiltcluster role to the fsicos2 role, we get the same effect, but faster.