icos.rdflog - make the status report work even in a readonly database
The status report about the rdflog database depends on a 'create function'
statement. That cannot be issued in the rdflog replicas since they're
readonly. If we instead add the functions when the database is initialized then
the function is part of the replication and can be used by the subscribers as
well (for their status report).