icos.password_env_file - new role
Generate passwords in docker compatible env files.

1. Store (if provided) or generate a password in the given file. It's stored in
   VAR=VALUE format.
2. The password is also read back and assigned to a variable.

The use-case is a scenario where a docker image requires a password during setup
- for example a database image. A second image might then need that to connect -
a web container connecting to a database for example.

But the password doesn't always need to be available outside the deployment. In
those cases this role can be used to generate the password during deployment and
share it between images.