sync-stiltweb-to-stilt should handle partly synced slots
It tries to sync three files from the new-style to the old-style
directory. If any of these three already existed it would skip the
remaining.

This patches changes this behaviour to always try all three files -
this makes it handle cases where the destination "slot" is broken (i.e
it has the Footprint file but not the RData).