icos.sshlogin - fix bug with user's home directory
sshlogin_src_home used the value of _src_user.home, which was only set after
executing a task that used sshlogin_src_home!

This recursive variable reference worked correctly for the first user, because
it used " | default(omit)", but for the second user, it used the value of
_src_user, which pointed to the previous user.

This meant that the second user created by sshlogin used the home directory of
the first user, which led later tasks astray since it tried writing to the
wrong directory.

Very confusing.